Added custom parameters to processing algorithms Added custom general error handler to catch pyQt5 errors (traceback) and to display error to user (TODO)
