+++ python/libopenimu/models/ProcessedData.py
@@ -7,9 +7,10 @@ from sqlalchemy.orm import relationship
 class ProcessedData(Base):
     __tablename__ = 'tabProcessedData'
     id_processed_data = Column(Integer, Sequence('id_processed_data_sequence'), primary_key=True, autoincrement=True)
-    id_data_processor = Column(Integer, nullable=False) #TODO: Relationship with data processor table...
+    id_data_processor = Column(Integer, nullable=False)
     name = Column(String, nullable=False)
     data = Column(BLOB, nullable=False)
+    params = Column(String, nullable=True)
     processed_time = Column(TIMESTAMP, nullable=False)
 
     # Relationships
@@ -17,5 +18,6 @@ class ProcessedData(Base):
 
     # Database rep (optional)
     def __repr__(self):
-        return "<ProcessedData(id_processed_data='%i', id_data_processor='%i', name='%s', data='%s')>" % (self.id_processed_data, self.id_data_processor, self.name, self.data)
+        return "<ProcessedData(id_processed_data='%i', id_data_processor='%i', name='%s', data='%s', params='%s')>" % \
+               (self.id_processed_data, self.id_data_processor, self.name, self.data, self.params)
