+++ python/resources/ui/MainWindow.ui
@@ -48,7 +48,8 @@ QPushButton[checkable=true]:!checked,QToolButton[checkable=true]:!checked{color:
 
 QLabel,QCheckBox{color:white;}
 
-QTreeWidget, QListWidget{background-color:rgba(150,150,150,50%);} 
+QListWidget{background-color:rgba(150,150,150,50%);}
+QTreeWidget{background-color:rgba(200,200,200,75%);} 
 
 QTreeWidget::item:hover,QTreeWidget::item:hover:selected {
 	border:none;
@@ -62,12 +63,7 @@ QTreeWidget::item:selected {
 }
 
 
-
-
-
-
-
-QLineEdit,QPlainTextEdit,QComboBox{background-color:rgba(226, 226, 226,90%);border-radius:5px;}
+QLineEdit,QPlainTextEdit,QComboBox, QSpinBox{background-color:rgba(226, 226, 226,90%);border-radius:5px;min-height:24px;}
 
 QListWidget#lstSensors,QTableWidget{background-color:rgba(200,200,200,70%);}
 
@@ -77,14 +73,48 @@ border: 2px solid darkgrey;
 background-color:grey;
 text-align: center;
 }
+
+QTabWidget::pane {
+    background-color: rgba(50,50,50,50%);
+    border-radius: 5px;
+    border: 2px solid rgba(128,128,140,50%);
+}
+
+QTabWidget::tab-bar {
+    left: 0px;
+}
+
+QTabBar::tab,QTableWidget::tab,QHeaderView::section {
+    background: qlineargradient(x1: 0, y1: 0, x2: 0, y2: 1, stop: 0  rgba(128,128,140,50%),  stop: 1.0  rgba(128,128,140,25%));
+    border: 2px solid  rgba(128,128,140,50%);
+    border-bottom: 0px;
+    border-top-left-radius: 4px;
+    border-top-right-radius: 4px;
+    padding: 2px;
+    min-height:25px;
+	color: white;
+}
+
+QTabBar::tab:selected, QTabBar::tab:hover {
+    background: qlineargradient(x1: 0, y1: 0, x2: 0, y2: 1, stop: 0  rgba(128,128,140,70%),  stop: 1.0  rgba(128,128,140,50%));
+}
+
+QTabBar::tab:selected {
+    border-color:rgba(128,128,140,60%);
+    border-bottom-color:rgba(128,128,140,60%);
+}
+
+QTabBar::tab:!selected {
+    margin-top: 2px;
+}
+
 QMainWindow::separator {
     background: grey;
     width: 5px; /* when vertical */
     height: 5px; /* when horizontal */
 }
 
-QLabel#lblVersion{color:black;}
-</string>
+QLabel#lblVersion{color:black;}</string>
   </property>
   <property name="documentMode">
    <bool>false</bool>
@@ -132,7 +162,7 @@ QLabel#lblVersion{color:black;}
    </property>
    <property name="minimumSize">
     <size>
-     <width>280</width>
+     <width>442</width>
      <height>180</height>
     </size>
    </property>
